feat(runtime): 完成坐标解密与云偏移闭环
- 接入严格校验的 T3 云布局与 Android HTTPS 传输
- 修复人物来源合并、屏内判定、相机缓存和健康字段链
- 提升人物绘制对比度并补齐多渲染后端恢复
- 从默认发布产物排除子弹追踪执行链和配置入口
- 补充模拟执行、运行时布局、认证及边界回归测试

+++ tests/projectile_speed_tests.cpp
@@ -43,7 +43,7 @@ void RunProjectileSpeedTests() {
 
     FakeMemory memory;
     memory.Put(weapon + 0x840, identity);
-    memory.Put(weapon + 0x838, std::uint64_t{830000001});
+    memory.Put(weapon + 0x838, std::uint32_t{830000001});
     memory.Put(weapon + 0x220, entries);
     memory.Put(weapon + 0x228, std::uint32_t{1});
     memory.Put(entries, field);
